SelectAction: Fix shift-select and ctrl-select.
Before r452 if you wanted to move all the selected objects you had to
shift-click on the selection and then drag.  r452 tried to remedy this
situation by clearing the selection if you didn't click on an already selected
object.  This made it impossible to add something to the selection by
shift-clicking it.

This change postpones replacing the selection until after the mouse released,
i.e. when we know whether the user wanted to select or move.