fix map flicker in TMSLayer drawing code when zooming out
I'm seeing the map flicker and black-out during zoom-out operations.  It
only seems to happen when I _start_ zoomed in to an area.  I believe
it's because there are no visible tiles in upper zoom levels and
displayZoomLevel gets set to 2 when the while loop goes a bit nuts.
Thus, no tiles get drawn since it's so far away from my actual zoom
level and it takes a moment for repaint to get called.

This patch will not bother moving displayZoomLevel unless we've
unquestionably _found_ some more visible tiles in another zoom level.