see #9516 - MapCSS: add basic support for a "contains" expression
The current syntax is `outer ∋ inner`, e.g., `way[building?] ∋ node[building?]`.

This implementation is known to be inefficient since, for every match on the right side, it checks all primitives of the dataset for a match.

Nevertheless, this implementation allows mappers to write specific MapCSS-based tests involving this spatial operator.