fix #17354: Relation list is not always updated
The problem was that `RelationListDialog.otherDatasetChange(AbstractDatasetChangedEvent event)` did nothing.
The dialog therefore misses events of type PRIMITIVE_FLAGS_CHANGED and if the flags of a relation are changed this should change the list.
When the dialog is build from scratch it looks at the relations with status complete and shows them.