see #6529: Deleting locally referenced objects in conflict resolution breaks data integrity - Correct calculation of commands when a selection of conflicts is resolved to "my" or "their" version
The old code computed a list of commands before executing any of them. This fails when the execution of a command changes the data that is used by following commands. I don't know if this fixes all problems. I expected that these conflicts also require a special order, but there is no code to sort them. The displayed list doesn't seem to be ordered.