see #10205: Strange align nodes in circle behavior - allow to define center node also when a single unclosed way is selected - add robustness and more unit tests for evaluation of valid selections - if only nodes and no way are selected, order the nodes using the angle to the agv. east-north position. This should produce a predictable result - if way(s) are selected the order the nodes is determined by the occurence in the way(s). Self-Intersecting polygons are rejected if no center node is given - don't throw InvalidSelection when selection is valid but no point was moved, let buildCommand() return null instead
With a selection that gives a center point and way(s) which are not even close to a circular shape the result might still be surprising. Sometimes the way nodes are arranged around the center node, sometimes they are moved so that a circle arc with nearly the same length is produced. The result changes significantly when the way nodes are also selected. Subject to further improvements.