Check current tile projection when querying WMS server.
Add same logic as in WMSEndpointTileSource to check, whether coordinates needs
to be switched or not. Before this change, switch was calculated only when
adding the layer, so changing map projection was ignored.

Also it was using JOSM map projection settings instead of the projection that
was chosen to query the server and this could result in replaced X and Y
coordinates.


Closes: #20509