fix #21919: NPE in MapImage due to a race condition
This adds a `synchronize` block where `disabledImgCache` is calculated.
The new `synchronize` block will vie for the same lock as that in `loadImage`.
There should not be any significant UI penalty from this, as the `loadImage`
synchronized block does not depend upon anything particularly expensive, like
a network call. It also helps that almost all the time where both blocks will
be in contention, the new `synchronize` block will win.