Fix #22073: Optimized regexes (string starts with, ends with, contains) do not work
This was caused by taking a portion of a regex (for example, `_name` from
`/_name$/`), and using that to match against keys. So `alt_name` would not
match, but `_name` would. All regex patterns should now be added to the
`remaining` field in `MapCSSRuleIndex`, and there now exists a test to ensure
we do not forget to add new non-regex types to the valid key type list.