Fix #22065: Mac users cannot deselect with ctrl
From the Mac OS HIG guidelines, "[as] much as possible,
avoid using the Control key as a modifier."

In this case, we were using ctrl to deselect elements (and
we disabled this on Mac), merge nodes, scale, and rotate.

All of those now use `⌘` (meta) instead of `ctrl` on Mac.