Fix #22106: Avoid allocations for point objects in QuadBuckets
Nodes do not keep their bboxes in memory (to reduce overall memory usage).
This does mean that every time someone searches a quadbucket for nodes, we
have to recreate the BBox for every node in a quadbucket leaf. This is a
significant number of allocations that are unnecessary (~80% of the memory
allocations from QuadBuckets#search for nodes is from the creation of
temporary bboxes). The remaining cost for the search is from List#grow calls.