MainApplication: Fix CORRECTNESS NP_NULL_PARAM_DEREF
This is mostly due to a false positive (the only way
for mainJOSM to be called with null is if
`new ProgramArguments` throws an exception, but we
exit the program in that case using Lifecycle.exitJosm).

In any case, moving `mainJOSM` into the `try-catch`
block should be marginally clearer, in so far as
it is easier to tell that we don't run `mainJOSM`
if we cannot parse the program arguments.