Fix #21630: Drop the "Riverbank" template (patch by zelonewolf)
TagInfo(1) shows that usage of waterway=riverbank has cratered over
the past year. As of writing, the total number of objects with
waterway=riverbank is less than 5,000 objects, and due to the significant
drop in usage, can be considered to be deprecated instead of
in use.

(1) taginfo:tags/waterway=riverbank#chronology