Fix #22183: NoClassDefFoundError: Could not initialize class org.openstreetmap.josm.actions.SessionSaveAction
This was caused by JOSM detecting that it had started in IPv6 mode, but no IPv6
network was available. When this occurs, JOSM immediately sets a preference key
and restarts so that users are not confused about a lack of network connectivity.

The JOSM restart action then proceeds to attempt to save any active layers, but
since this occurs during startup, there are no active layers and the save actions
have not yet been initialized. The save actions cannot be initialized before
`contentPanePrivate` has been initialized, but the IPv6 call is done prior to
that happening, which may then trigger a JOSM restart, which calls `exitJosm`,
which then calls the save layers actions.