Fix #22606: Images with no exif position or no set position were treated as if they had no position
This is incorrect behavior, since the paint code only cares about the return
value of IImageEntry#getPos, and invalid geoimages should be those that can
not be shown in the mapview.

This also fixes an issue where, when the layer has no images with positions, the
first image would be displayed, but not selected in the data.