Fix #22744, see #20768: Fix NPE when switching to authentication methods
The NPE occurs because the UI state and the saved preferences state does not
match. This is solved by saving the state to preferences and then reloading it.

This also fixes an issue where the `http2` plugin would prevent the OAuth
workflow from working, due to differing default behaviors (see #22730).