Fix #22783: Presets menu should not sort the first action block
The specific problem was that when the preset menu was sorted, the order of
the  Preset menu (prior to the first separator) was `Search preset...`,
`Preset preferences...`, `Search for objects by preset...` instead of
`Search preset...`, `Search for objects by preset...`, `Preset Preferences...`

The fix for this was changing the TaggingPresetMenu.PresetTextComparator.compare
method to have the same behavior for the ordered presets. A non-regression
test was also added.