See #22767, fix failing unit tests
The fix for #22767 included a new unit test, which caused a cascading failure.
The first step of the fix included removing the destroyed dialog from the
MapFrame. Unfortunately, `MapFrame#removeToggleDialog` did not properly
remove the ToggleDialog -- it called `DialogsPanel#remove(Component)` with
the ToggleDialog as an argument. Since the DialogsPanel never adds the
ToggleDialog as a direct subcomponent, this was effectively a no-op.