Fix #22726: DownloadOsmChange task should not always get the history for each unknown object
In many cases, the latest version of an object will also be the version we will
get when iterating through all of its history. This will always be the case for
objects with only one version. As such, in most cases, getting the latest version
of an object will significantly reduce the number of individual history calls.

In a test, the amount of time taken to load an OSC went from 105 seconds to 14
seconds. This is an 86% improvement in terms of time taken.