Fix #23001: ClassCastException in GpxDrawHelper#calculateColors
This occurs when a working GPX file with HDOP information is converted to an OSM
Data Layer and back to a GPX Data Layer. This is fixed by (a) converting casts
for `get` operations to `Number` instead of `Float`/`Integer` and (b) actually
putting the number in the attribute map instead of a string.