Fix #23015, see #2483: Right click (map move) + scale/rotate modes should not interfere with each other
The workaround added for #2483 (r1634, Java 5) does not appear to be needed
anymore (tested with a Magic Mouse in one-button mode under Java 8). Removing
the workaround fixes the behavior whereby a right-click will cause the
scale/rotate mode to be enabled.

To fix extraneous commands being generated during a map pan when scale/rotate
mode is enabled, we check to see if the right mouse button is down in
`updateCommandWhileDragging`.