See #22652, r18817: Stop parsing gpx files when point elements have invalid coordinates
This adds back the wpt check that was temporarily removed in r18818 to get back
to the previous "known good" state and additionally adds checks for rtept and
trkpt elements.

This also removes a change to build.xml that was accidentally committed in r18818.
The specific change (adding `Automatic-Module-Name` to the JOSM manifest) will
probably be done in a separate commit in the future.