Fix #17052: Allow plugins to save state to session file
The primary feature request was for the TODO plugin to save the list elements for
a future session.

This allows plugins to register via `ServiceLoader` classes which need to be
called to save or restore their state.

In addition, this fixes an ordering issue with tests whereby the OsmApi cache
would be cleared, but the FakeOsmApi class would not recache itself when called.