Fix #23203: Simplify Way dialog can cause a deadlock
This is caused by a write lock being acquired in `SimplifyWayAction.actionPerformed`
in the EDT and then a read lock is attempted to be acquired in a styled-map-renderer
thread. The styled-map-renderer thread is waited on by the EDT, which causes the
deadlock. The fix is removing the dataset write lock in `SimplifyWayAction.actionPerformed`
which was added back in 2011. An analysis of the code shows that the commands generated
lock the dataset themselves, so the dataset lock in `SimplifyWayAction.actionPerformed`
should no longer be necessary.

The deadlock does not happen reliably with only the JOSM default paintstyle; adding
additional paintstyles was required for the initial analysis.