See #16567: Update to JUnit 5
This adds the @TaggingPresets annotation used by r18866 (see #23196). The
annotation tries to only initialize the presets when the current preset list does
not have the same hashCode as the default preset list. In order for this to work,
`TaggingPresets#getTaggingPresets` had to return a `List` (the method contract
has not changed), which calculates the hashCode in a deterministic manner.