Fix #23341: Don't set paintstyles to nothing when paintstyles haven't been modified by the user
The original fix (r18907) was a short-term fix that reset paintstyles to the
default if the list was empty. This is not ideal, since people who have
deliberately made their paintstyle list empty now have to clear it every time
JOSM starts.

This patch fixes that by only attempting to remove potlatch2 if the user has
previously modified the painstyle list. The previous patch was necessary to get
most users back to a "known good" state, especially those that have never used
the style settings.