Fix #23687: Freeze when opening `Add Tag` dialog
This occurs when the following are true:
* The EDT is in a blocking dialog where a DataSet write lock is held
* The EDT decides to repaint the mapview
* The style caching occurs on a different thread

Why is this a problem?
The style caching waits for a read lock in a different thread, which will never
happen since the EDT has a write lock. The try-catch block that checks to see if
the current thread can get a read lock does not work on a thread that already has
a write lock, since that thread will always be able to obtain a read lock.

There are two solutions to this:
1. Change the style caching code such that it times out after some period of time
   and then retries in the EDT (to avoid recurrent timeouts)
2. Change TagEditHelper such that the blocking dialog is not in a block which
   holds a DataSet write lock
 
Option 2 is better, since it fixes the root cause; option 1 may need to be done
in the future, if only to allow for bug reports.