Fix #23804: xdg-open can return a `0` exit code when a browser doesn't open
This is alleviated by switching the default order of `browser.unix` to be
`#DESTKOP` then `xdg-open`. `#DESKTOP` is our magic value for using built-in
Java methods for opening a browser.

It ''appears'' that there is a race condition somewhere for the Brave browser.
I ''suspect'' that the process returned by the Runtime.exec and xdg-open isn't
actually complete when it says it is, and so the process is cleaned up prior to
Brave actually launching.