Fix #23860: Duplicate key+value in preset causes an ISE in TagChecker
This fixes the actual issue in TagChecker, but also adds a sanity check to
`TaggingPresetPreferenceTestIT` since it is ''usually'' unintended to have
duplicate key/values (and it is always a problem if they are different and are
`Key` items).

The fix for `TagChecker` is just keeping whatever `value` is last. Not ideal, but
it should work 99% of the time since an object won't match the preset if we have
`highway=footway` and `highway=footway2` as `Key` objects.