Fix #23837: Set `-XX:MaxRAMPercentage` to >25% of system memory
This does not currently work for WebStart, see
https://github.com/AdoptOpenJDK/IcedTea-Web/pull/954 .

As noted in the ticket, all supported MS Windows and Apple macOS machines have
4G+ of RAM. Linux machines don't have a baseline, and some have 1G or less. Where
we detect that a machine has <8G of memory in the linux start scripts, we don't
add the JVM parameters.

As a reminder to everyone reading the changelog: just because we ''ask'' for the
ability to use 75% of the system memory doesn't mean we will ''use'' all of that
memory. The JVM ''will'' free up unused memory for the rest of the system.

We additionally set the minimum memory allocation to 256m; while this ''can'' be
lower, that is for minimal amounts of data loaded. A clean JOSM install will use
99mb, for example. But once we add a minimal amount of data and a background
layer, static memory usage goes up to 161mb.