Add JUnit5 tags to integration tests, functional tests, and performance tests
This should allow contributors to filter out integration tests inside the standard
test directory. Of specific note, this should reduce test runtime for contributors
by a non-trivial amount if contributors filter out the tagged tests.

For consistencies’ sake, tests in `functional` and `performance` were also
annotated. As such, if we ever merge the test directories, this will let us
filter which tests are run in CI.

Note that many functional tests don't have the annotation. This is because they
are ''manual'' tests.