See #24075: Reduce memory allocations for TaggingPresetItem#matches
Make PMD happy again.

As noted in the comments for the for loop in question, this is a fairly
significant performance optimization. The for-each loop that PMD prefers has
significant penalties on hot code sections for `ArrayList` objects, which we use
extensively.

TBH, the JVM should probably do this optimization for (at minimum) ArrayList
objects.